this looks not bad as a first avenue
