+++ bionc/math_interface/using_numpy/__init__.py
@@ -1,0 +1,1 @@
+from .natural_coordinates import SegmentNaturalCoordinates
