+++ bionc/model_computations/natural_marker.py
@@ -5,8 +5,9 @@ import numpy as np
 from .biomechanical_model import BiomechanicalModel
 from ..model_creation.protocols import Data
 from ..utils.interpolation_matrix import interpolate_natural_vector, to_natural_vector
-from ..utils.natural_coordinates import SegmentNaturalCoordinates
 
+# from ..utils.natural_coordinates import SegmentNaturalCoordinates
+from ..math_interface.internal import SegmentNaturalCoordinates
 
 # todo: need a list of markers MarkerList
