+++ bionc/utils/__init__.py
@@ -1,4 +1,6 @@
 from .homogenous_transform import HomogeneousTransform
-from .natural_coordinates import SegmentNaturalCoordinates, NaturalCoordinates
+from .natural_coordinates import (
+    # SegmentNaturalCoordinates, \
+    NaturalCoordinates)
 from .natural_velocities import SegmentNaturalVelocities, NaturalVelocities
 from .natural_accelerations import SegmentNaturalAccelerations, NaturalAccelerations
